requirements.txt
numpy==1.20.3
requests-oauthlib==1.3.0
matplotlib==3.4.2
sentinelhub==3.3.0
pypng==0.0.20
requests==2.25.1
oauthlib==3.1.1
Pillow==8.2.0
flask==2.0.1